+++ flink-table/flink-table-planner/src/main/scala/org/apache/flink/table/planner/plan/nodes/physical/stream/StreamPhysicalIntervalJoin.scala
@@ -27,7 +27,7 @@ import org.apache.flink.table.planner.plan.nodes.exec.{ExecNode, InputProperty}
 import org.apache.flink.table.planner.plan.nodes.physical.common.CommonPhysicalJoin
 import org.apache.flink.table.planner.plan.utils.PythonUtil.containsPythonCall
 import org.apache.flink.table.planner.plan.utils.RelExplainUtil.preferExpressionFormat
-import org.apache.flink.table.planner.utils.ShortcutUtils
+import org.apache.flink.table.planner.utils.ShortcutUtils.unwrapConfig
 
 import org.apache.calcite.plan._
 import org.apache.calcite.rel.core.{Join, JoinRelType}
@@ -93,7 +93,7 @@ class StreamPhysicalIntervalJoin(
 
   override def translateToExecNode(): ExecNode[_] = {
     new StreamExecIntervalJoin(
-      ShortcutUtils.unwrapConfig(this),
+      unwrapConfig(this),
         new IntervalJoinSpec(joinSpec, windowBounds),
         InputProperty.DEFAULT,
         InputProperty.DEFAULT,
