+++ flink-table/flink-table-planner/src/main/scala/org/apache/flink/table/planner/plan/nodes/physical/stream/StreamPhysicalWatermarkAssigner.scala
@@ -23,7 +23,7 @@ import org.apache.flink.table.planner.plan.nodes.calcite.WatermarkAssigner
 import org.apache.flink.table.planner.plan.nodes.exec.stream.StreamExecWatermarkAssigner
 import org.apache.flink.table.planner.plan.nodes.exec.{ExecNode, InputProperty}
 import org.apache.flink.table.planner.plan.utils.RelExplainUtil.preferExpressionFormat
-import org.apache.flink.table.planner.utils.ShortcutUtils
+import org.apache.flink.table.planner.utils.ShortcutUtils.unwrapConfig
 
 import org.apache.calcite.plan.{RelOptCluster, RelTraitSet}
 import org.apache.calcite.rel.{RelNode, RelWriter}
@@ -71,7 +71,7 @@ class StreamPhysicalWatermarkAssigner(
 
   override def translateToExecNode(): ExecNode[_] = {
     new StreamExecWatermarkAssigner(
-      ShortcutUtils.unwrapConfig(this),
+      unwrapConfig(this),
       watermarkExpr,
       rowtimeFieldIndex,
       InputProperty.DEFAULT,
